Use correct type for uframe and utool in robot_pose

+++ src/include/fanuc-ethernet.hpp
@@ -16,8 +16,8 @@ using namespace eipScanner::utils;
 
 namespace fanuc_ethernet {
     struct robot_pose {
-        int utool;
-        int uframe;
+        short utool;
+        short uframe;
         float x;
         float y;
         float z;
